refactor: styled -> dynamic
