+++ src/pages/[post]/_components/Content/style/Footnotes.ts
@@ -1,4 +1,4 @@
-import { css } from '@/ui/styled';
+import { css } from '@/ui/styled/dynamic';
 
 const Footnotes = css`
   [data-footnotes] {
